+++ src/main/java/com/mitchellbosecke/pebble/attributes/ArrayResolver.java
@@ -1,6 +1,7 @@
 package com.mitchellbosecke.pebble.attributes;
 
 import com.mitchellbosecke.pebble.error.AttributeNotFoundException;
+import com.mitchellbosecke.pebble.template.EvaluationContextImpl;
 
 import java.lang.reflect.Array;
 
@@ -14,14 +15,14 @@ public class ArrayResolver implements AttributeResolver {
   public ResolvedAttribute resolve(Object instance,
                                    Object attributeNameValue,
                                    Object[] argumentValues,
-                                   boolean isStrictVariables,
+                                   EvaluationContextImpl context,
                                    String filename,
                                    int lineNumber) {
     String attributeName = String.valueOf(attributeNameValue);
     int index = this.getIndex(attributeName);
     int length = Array.getLength(instance);
     if (index < 0 || index >= length) {
-      if (isStrictVariables) {
+      if (context.isStrictVariables()) {
         throw new AttributeNotFoundException(null,
                 "Index out of bounds while accessing array with strict variables on.",
                 attributeName, lineNumber, filename);