+++ src/main/java/com/mitchellbosecke/pebble/attributes/AttributeResolver.java
@@ -1,10 +1,12 @@
 package com.mitchellbosecke.pebble.attributes;
 
+import com.mitchellbosecke.pebble.template.EvaluationContextImpl;
+
 public interface AttributeResolver {
   ResolvedAttribute resolve(Object instance,
                             Object attributeNameValue,
                             Object[] argumentValues,
-                            boolean isStrictVariables,
+                            EvaluationContextImpl context,
                             String filename,
                             int lineNumber);
 }