+++ CHANGELOG.md
@@ -18,6 +18,7 @@
 - Make "loop.length" and "loop.revindex" be lazy evaluated (#279)
 - Fixed security issue which allowed to execute shell command (by having access to Java's Class object) (#329)
 - Throw ParserException when endBlock not found (#308)
+- For tag : add enumeration support (#292)
 
 ## v2.4.0 (2017-06-04)
 - Add arrays support for iterable test (#254)