+++ CHANGELOG.md
@@ -19,6 +19,7 @@
 - Fixed security issue which allowed to execute shell command (by having access to Java's Class object) (#329)
 - Throw ParserException when endBlock not found (#308)
 - For tag : add enumeration support (#292)
+- Fix operator precedence (#290)
 
 ## v2.4.0 (2017-06-04)
 - Add arrays support for iterable test (#254)