+++ CHANGELOG.md
@@ -11,7 +11,7 @@
 - Add the possibility to extend attribute access (#258)
 - Remove DynamicAttributeProvider
 
-## v2.5.0 (2018-xx-xx)
+## v2.5.0 (2018-05-02)
 - Handle all Iterables and Arrays in JoinFilter (#273)
 - Fix NPE in GetAttributeExpression (#299)
 - Remove redundant null check before calling operator "instanceof" in OperatorUtils. (#298)